# Speedup feature selection function